refactor(editor): nested list item model 분리
